+++ src/core/HorizontalTabs/HorizontalTabs.test.tsx
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 import { fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
-
+import { HorizontalTab } from './HorizontalTab';
 import { HorizontalTabs, HorizontalTabsProps } from './HorizontalTabs';
 
 const renderComponent = (
@@ -12,43 +12,49 @@ const renderComponent = (
   initialChildren?: React.ReactNode,
 ) => {
   const defaultProps: HorizontalTabsProps = {
-    labels: ['item1', 'item2', 'item3'],
+    labels: [
+      <HorizontalTab key={1} label='Label 1' />,
+      <HorizontalTab key={2} label='Label 2' />,
+      <HorizontalTab key={3} label='Label 3' />,
+    ],
   };
   const props = { ...defaultProps, ...initialProps };
   const children = initialChildren ?? 'Test content';
   return render(<HorizontalTabs {...props}>{children}</HorizontalTabs>);
 };
 
-it('should render tags', () => {
+it('should render tabs', () => {
   const { container } = renderComponent();
 
-  const tabContainer = container.querySelector('ul') as HTMLElement;
+  expect(container.querySelector('.iui-tabs-wrapper')).toBeTruthy();
+
+  const tabContainer = container.querySelector('.iui-tabs') as HTMLElement;
   expect(tabContainer).toBeTruthy();
   expect(tabContainer.className).toEqual('iui-tabs iui-default');
-  expect(tabContainer.querySelectorAll('button').length).toBe(3);
+  expect(tabContainer.querySelectorAll('.iui-tab').length).toBe(3);
   screen.getByText('Test content');
 });
 
-it('should render borderless tags', () => {
+it('should render borderless tabs', () => {
   const { container } = renderComponent({ type: 'borderless' });
 
-  const tabContainer = container.querySelector('ul') as HTMLElement;
+  const tabContainer = container.querySelector('.iui-tabs') as HTMLElement;
   expect(tabContainer).toBeTruthy();
   expect(tabContainer.className).toContain('iui-tabs iui-borderless');
 });
 
-it('should render pill tags', () => {
+it('should render pill tabs', () => {
   const { container } = renderComponent({ type: 'pill' });
 
-  const tabContainer = container.querySelector('ul') as HTMLElement;
+  const tabContainer = container.querySelector('.iui-tabs') as HTMLElement;
   expect(tabContainer).toBeTruthy();
   expect(tabContainer.className).toContain('iui-tabs iui-pill');
 });
 
-it('should render green tags', () => {
+it('should render green tabs', () => {
   const { container } = renderComponent({ color: 'green' });
 
-  const tabContainer = container.querySelector('ul') as HTMLElement;
+  const tabContainer = container.querySelector('.iui-tabs') as HTMLElement;
   expect(tabContainer).toBeTruthy();
   expect(tabContainer.className).toContain('iui-green');
 });
@@ -57,7 +63,7 @@ it('should call onTabSelected when switching tabs', () => {
   const onTabSelected = jest.fn();
   const { container } = renderComponent({ onTabSelected });
 
-  const tabs = container.querySelectorAll('button');
+  const tabs = container.querySelectorAll('.iui-tab');
   expect(tabs.length).toBe(3);
   fireEvent.click(tabs[2]);
   expect(onTabSelected).toHaveBeenCalledWith(2);
@@ -66,7 +72,7 @@ it('should call onTabSelected when switching tabs', () => {
 it('should set active tab', () => {
   const { container } = renderComponent({ activeIndex: 2 });
 
-  const tabs = container.querySelectorAll('button');
+  const tabs = container.querySelectorAll('.iui-tab');
   expect(tabs.length).toBe(3);
   expect(tabs[0].className).not.toContain('iui-tab iui-active');
   expect(tabs[1].className).not.toContain('iui-tab iui-active');
@@ -76,11 +82,43 @@ it('should set active tab', () => {
 it('should not fail with invalid active tab and set the first one', () => {
   const { container } = renderComponent({ activeIndex: 100 });
 
-  const tabs = container.querySelectorAll('button');
+  const tabs = container.querySelectorAll('.iui-tab');
   expect(tabs.length).toBe(3);
   expect(tabs[0].className).toContain('iui-tab iui-active');
   expect(tabs[1].className).not.toContain('iui-tab iui-active');
   expect(tabs[2].className).not.toContain('iui-tab iui-active');
+});
+
+it('should render strings in HorizontalTab child component', () => {
+  const { container } = renderComponent({
+    labels: ['item0', 'item1', 'item2'],
+  });
+
+  const tabs = container.querySelectorAll('.iui-tab');
+  expect(tabs.length).toBe(3);
+  tabs.forEach((tab, index) => {
+    const label = tab.querySelector('.iui-tab-label') as HTMLElement;
+    expect(label).toBeTruthy();
+    expect(label.firstElementChild?.textContent).toEqual(`item${index}`);
+  });
+});
+
+it('should add .iui-large if HorizontalTab has sublabel', () => {
+  const { container } = renderComponent({
+    labels: [
+      <HorizontalTab key={0} label='item0' sublabel='Sublabel0' />,
+      <HorizontalTab key={1} label='item1' sublabel='Sublabel1' />,
+      <HorizontalTab key={2} label='item2' sublabel='Sublabel2' />,
+    ],
+  });
+  expect(container.querySelector('.iui-tabs.iui-large')).toBeTruthy();
+
+  const tabs = container.querySelectorAll('.iui-tab');
+  expect(tabs.length).toBe(3);
+  tabs.forEach((tab, index) => {
+    const label = tab.querySelector('.iui-tab-label') as HTMLElement;
+    expect(label.textContent).toEqual(`item${index}Sublabel${index}`);
+  });
 });
 
 it('should add custom classnames', () => {